Changed proxy servlet name to ProxyServlet.
